Revert a few commits and release 1.2.0
Commits fd92787 and d7ac798 won't be available in the middleware release
of 1.2.0 because they happened after we were forced onto clojure 1.7 so
they have been reverted to cut this release.

The purpose of this release is to get as many features out as possible
to the users who can't upgrade to clojure 1.7 or later.

1.2.0 will be the final release compatible with clojure versions below
1.7, making the next version 2.0.0.